:recycle: Refactor : header component Folder
