+++ src/pages/changePassword/ChangePassword.tsx
@@ -10,7 +10,7 @@ import Swal from 'sweetalert2';
 
 import { ChangeOfPassword } from '../../api/userApi';
 import BackPageIcon from '../../components/BackPageIcon';
-import Header from '../../components/Header';
+import Header from '../../components/header/Header';
 import RegexHelper from '../../helper/RegexHelper';
 import useError from '../../hooks/useError';
 import useLoginInfo from '../../hooks/useLoginInfo';
