+++ src/pages/editThread/EditThread.tsx
@@ -9,7 +9,7 @@ import Swal from 'sweetalert2';
 
 import { fetchMainTextThread, updateThread } from '../../api/threadApi';
 import BackPageIcon from '../../components/BackPageIcon';
-import Header from '../../components/Header';
+import Header from '../../components/header/Header';
 import Spinner from '../../components/Spinner';
 import RegexHelper from '../../helper/RegexHelper';
 import useError from '../../hooks/useError';
