+++ src/pages/notfound/NotFound.tsx
@@ -3,7 +3,7 @@ import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { useNavigate } from 'react-router-dom';
 import styled from 'styled-components';
 
-import Header from '../../components/Header';
+import Header from '../../components/header/Header';
 import Meta from '../../Meta';
 
 const Wrapper = styled.section`
